chore(release): add script to create deb and rpm packages
Introduce scripts/build-packages/build-packages script.

This script also handles the creation of ggshield wheel and zipapp,
which were previously directly implemented in the GitHub workflow.
Having these steps in the script makes it easier to test the build from
outside the CI.

The script uses nfpm to build deb and rpm packages.

The setup.cfg change is there to fix a deprecation warning which appears
when running `python setup.py --version`.